Fix ALSA output to work with 24bps streams.
Seems its 24-bit integer option isn't well supported
so it's better to switch to floats in that case.

+++ src/output/alsa.h
@@ -31,7 +31,7 @@ typedef struct {
     union {
         int8_t *int8;
         int16_t *int16;
-        int32_t *int24;
+        float *float32;
     } buffer;
 
     PyObject *framelist_type;
